# perf: ~90% faster `get_cached_value` from Redis cache