Don't convert `_dict` to `Document` based on new keyword argument `_allow_dict` in `get_cached_doc`
(prefixed with underscore to prevent clash with actual fieldname)

This change impacts performance only when retrieving document from Redis Cache, but this is almost all the time because `document_cache` is only for specific request. Local cache code was commented for below tests.

### Before

```python
In [2]: %timeit frappe.get_cached_value("Sales Invoice", "SINV-21-00070", "name")
214 µs ± 779 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

### After (99% improvement for Sales Invoice)

```python
In [2]: %timeit frappe.get_cached_value("Sales Invoice", "SINV-21-00070", "name")
1.39 µs ± 17.2 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```



---

Minor new feature:  `fieldname` in `get_cached_value` defaults to `name`, just like `db.get_value`